ecommerce_rest/requirements.txt
Django==3.1.7
django_simple_history==2.11.0
djangorestframework==3.13.1
djangorestframework_simplejwt==5.0.0
drf_yasg==1.20.0
Faker==13.3.4
